FieldConvert/ProcessInterpPointDataToFld: Added ability to interpolate 3D fields.
This adds a simplified/memory-optimized version of the Shepard algorithm to enable
interpolatin of 3D data. For 1D-data, the original quadratic/linear algorithm are chosen
automatically. In addition, the weights are stored for each point, which might be useful
for e.g running the same interpolation multiple times in future implementations.