Enable timedependent file function type.
This adds a TIMEDEPENDENT argument to Function type File and enables the use of
a time depenmdent file name for boundary conditions, auxiliary fields etc.

The implementation relies on boost::format, where the given filename is
used as format string and the step time as argument. The full boost::format
syntax is supported. An error is thown when the corresponding file is not
found, so in order to only supply a new file every Nth timespep, one might
use appropriately named symlinks.

Example:

The function
<FUNCTION NAME="Baseflow">
    <F VAR="U0" TIMEDEPENDENT="1" FILE="U0_%14.8R.fld"/>
</FUNCTION>

will look for files named U0_basegrid_1.00000000E-05.fld,
U0_basegrid_2.00000000E-05.fld and so on.