Make the tests run on Cray supercomputers - see issue #154
I've modified the Tester driver program to be configured by CMake. The
MPIEXEC program and its flag to specify the number of processes is now
given through CMake and a NEKTAR_TEST_FORCEMPIEXEC option added. If
the compiler has MPI_BUILTIN (as on Cray machines), then the minimum
necessary options that would have been set by FindMPI.cmake are set.