APE: Retain original APE4 formulation by Ewert and Schröder
The equations that were implemented before had the speed of sound moved into the
divergence term of the p-equation. This implicates an additional source term
that depends on the gradient of c, which was not present. Implementing this
source term has proven to produce instabilities at the element boundaries due to
the unsteady gradients. Accordingly, this implementation modifies the flux
formulation and retains Ewert and Schröders original formulation of the APE 1/4
equations LHS.

This fixes the behavior of the equations at e.g. density or pressure gradients.