This is round 2 of the Time Integration refactoring.  I like this arrangement better.  I have removed from the "Integrators" (that I added in the previous commit) and just refer to them as schemes now. A scheme (also called a "master scheme") can have a number of phases (or sub-schemes).  Each sub-scheme uses a "scheme" to set up the data for that phase.
I have put the code that creates the (full/master) scheme in the corresponding file, eg:

AdamsBashforthOrder2TimeIntegrationScheme.h  (see the constructor)

Additionally, the code that sets the data for the sub-scheme/phase that corresponds to
the master scheme is also in the same file (see the SetupSchemeData() method).
The data for each phase of the scheme is contained in the TimeIntegrationSchemeData object.

This code refactoring removes the giant switch statement that was used to initialize
the schemes.  It also removes the use the Memory Manager that was returning shared
pointers to each sub-scheme - which, while it worked, would not have allowed multiple
schemes to be used in the same executable.

I have also moved the TimeIntegrationSolution object into its own file.

--------------

Note, to create an integration scheme, do something like the following (using the factory):

    LibUtilities::TimeIntegrationSchemeSharedPtr tiScheme;

    TimeIntegrationSchemeFactory & factory = LibUtilities::GetTimeIntegrationSchemeFactory();

    tiScheme = factory.CreateInstance( "IMEXOrder1" );

Note, as I go to commit this, it occurs to me that instead of passing a string to CreateInstance(),
we should be passing the enum type.  I'll update this soon.

--------------

There are currently 3 failing tests (of 459):

	275 - IncNavierStokesSolver_Prism_channel_varP (Failed)
	326 - LibUtilities_TimeIntegrationDemoIMEXOrder2 (Failed)
	341 - StdRegions_StdProject2D_Quad_Mod_P6_Q7 (Failed)

Note, "326 - LibUtilities_TimeIntegrationDemoIMEXOrder2" is failing because IMEXOrder2 is done
differently between the TimeIntegration/ directory and the original TimeIntegrationDemo code
(which "re-implemented" the creation of the schemes).  (IMEXOrder2 should use IMEXdirk_2_3_2 as
the 1st phase, but it is currently set to IMEXOrder1 instead.  Once a definitive determination
on this is made, this needs to be updated one way or the other.)

It is unclear to me why the other 2 are failing (and the errors are very minuscule).  For example:

275:   Expected: 0 +/- 1e-12
275:   Result:   1.27742e-12

Before this branch gets merged in, we need to track down the reason for these differences.