Based on feedback from Dave M and Chris C, made the following updates:
- Returned to using a string name (instead of enum) for the factory creation of the scheme.
- Moved the scheme factory registrations out of the TimeIntegrationScheme.cpp file
    and into the separate SchemeInitializor.cpp file.  This was a compromise
    from creating a separate .cpp file for every specific scheme file to do the registration.
- Grouped the related Time Integrators into shared .h files.  (For example, all
    Runge Kutta type Time Integrators are now defined in the RungeKuttaTimeIntegrationSchemes.h
    file.

As this is mostly a "cosmetic" update, all Ctests (that were passing) are still passing.
The ones that are not passing are related to the difference in implementations of time
integrators between the wrappers (that are being replaced in this branch) and the time
integrator demo.  See details as specified in my previous commit:

   commit e66dd8daaa7bb666c4330e17f2a9d01d092a69e1
   Author: J. Davison de St. Germain <dav@sci.utah.edu>
   Date:   Sun May 19 22:01:22 2019 -0600

This still needs to be addressed and I believe is on Chris' to-do list.