Fix for comparing Nodes and Floats and Improved NekMesh peralign
Updated implementation for comparing nodes and floats:
    - Rely on boost for comparing floating point values using a
      relative distance.
    - Add alternative routine that compare Nodes and Floats using
      absolute tolerance.

Improved peralign command for NekMesh:
    - Allow the user to use both relative and absoltue tolerance
      for aligning the mesh at the periodic boundaries
    - By default, the code only uses the relative tolerance, as it
      did before.
    - Improved readability of code by reorganizing some if-statements
    - Changed help commands by improving formatting of text

Previously, the code featured our own implementation for comparing two floating
point values based on a relative distance. Given the many pitfalls of comparing
floats, there is always the risk that we overlook some special case. To improve
robustness, we now rely on the boost implementation to compare floats using a
relative distance. The boost routines will give identical results to previous
implementation, and hopefully also improve robustness and correctness for some
special cases that we overlooked before.

When a relative comparison is meaningless, such as when both floats are very
close to zero, we now also have a routine for comparing the distance using an
absolute tolerance (called IsRealClose + IsNodeClose). This routine is utilized
in the updated peralign command for NekMesh.