New Feature for Testing if Warnings Appear
Previously, the code could only test if a particular pattern was
written to the output stream, such as L2 error within a given tolerance.
This testing capability was further designed such that all patterns
defined by the user had to appear in the output stream for the test
to pass.

A warning only appears in the output stream if the code caught
something wrong, and the test should thus only fail if the user
says that "any of the following error/warnings" are not accepted.

Using the previous testing features to "fail if any of these patterns
are present" is not possible. Simply defining a derived class of the
previous test feature, which negates its result, is also not possible.
That is, suppose we specify 3 different Warning messages that we wish
to fail if present. If the previous test finds one of them, but not the
other two, it returns false. Negating false becomes true, and the test
passed, even though we wish the opposite. In fact, only if all
warning messages are present, would we (correctly) fail the test.

To remedy the above problem, a new derived class from Metric, called
MetricNoWarning is being implemented. It fails the test if it finds
any of the regex patterns the user specifies. It can thus be used to
scan the output for unwanted messages, and fail if they are found.