New Feature to check if Warnings appear
Added test to CMakeList.txt and checked that it compiles.

Did initial tests to see if it works. At present, the regex matching
and logic appears to work. I.e., it can find a regexp in the output
and match it against a pre-defined group (subexpression). If all sub-
expressions match, for any of the match tags, the test returns false
as it should.