New Feature to check if Warnings appear
Updated the new test routine with ERROR handlers to catch wrong input from user
throught the ASSERTL0() routine.

Added loop over both standard output and error output to catch potential error
put in the error stream as opposed to the standard output stream.

I have manually tested that all ASSERTL0 are working correctly by deliverately
writing an invalid .tst file and then running the test.

I updated the peralign_rot_cyl testcase to include the new NoWarning test, and
updated the command that is executed when this test is run to ensure that no
warnings appear. Before, this test raised a waring that was not caugt, now it
will be caugt if the code raises a warning.