Fixed Indentation and changed ASSERTL0(false,) to NEKERROR()
Found several places where the code uses ASSERTL0(false, "message") to report an error.
As suggested by Chris and the code guidelines, it is preferred to use NEKERROR()
in these cases. After going through changes in MR!1212, I found a couple of these
occurences and decided to change to NEKERROR. The changes have no effect on the code
since ASSERTL0(false,) calls NEKERROR().