Fixing Aliases for Nested Arrays (TensorOfArrayND)
Removed all occurances of TensorOfArray1D, and replaced with
Array<OneD, NekDouble> to stay consisten with the rest of the code.
In cases the TensorOfArray was const declared, it was replaced with
const Array<OneD, const NekDouble>. This double use of const is used
extensively in the code when 1D arrays are passed.

Removed duplicate aliases for nested arrays in SharedArray.hpp. Also
removed aliases declared in this file that were not used anywhere in
the code.

Next, the TensorOfArray2D alias will be removed.